fix/test: refactor caused test failures
